Fixed few typos in entity-list.ts
I loved your portfolio and noticed on the very first dialog the word "intereact". At first I lowkey wondered if it was a play on words because you use Svelte (=/= React), but then I decided to edit the 2 mistakes I saw and help you in your project.

Congratulations for your portfolio and for creating Skeleton UI as well! :)